Reap child processes after each of the ssh tests in case of timeouts
