+++ tests/conftest.py
@@ -5,7 +5,6 @@
 import logging
 import os
 import pathlib
-import pprint
 import re
 import shutil
 import stat
@@ -16,7 +15,6 @@ from unittest import TestCase  # pylint: disable=blacklisted-module
 import _pytest.logging
 import _pytest.skipping
 import more_itertools
-import psutil
 import pytest
 
 import salt
@@ -1299,7 +1297,6 @@ def salt_call_cli(salt_minion_factory):
 
 @pytest.fixture(scope="session", autouse=True)
 def bridge_pytest_and_runtests(
-    reap_stray_processes,
     salt_factories,
     salt_syndic_master_factory,
     salt_syndic_factory,
@@ -1336,6 +1333,8 @@ def bridge_pytest_and_runtests(
         salt_syndic_factory.config["conf_file"]
     )
     RUNTIME_VARS.TMP_SSH_CONF_DIR = str(sshd_config_dir)
+    with reap_stray_processes():
+        yield
 
 
 @pytest.fixture(scope="session")
@@ -1635,46 +1634,6 @@ def from_filenames_collection_modifyitems(config, items):
 
 
 # ----- Custom Fixtures --------------------------------------------------------------------------------------------->
-@pytest.fixture(scope="session")
-def reap_stray_processes():
-    # Run tests
-    yield
-
-    children = psutil.Process(os.getpid()).children(recursive=True)
-    if not children:
-        log.info("No astray processes found")
-        return
-
-    def on_terminate(proc):
-        log.debug("Process %s terminated with exit code %s", proc, proc.returncode)
-
-    if children:
-        # Reverse the order, sublings first, parents after
-        children.reverse()
-        log.warning(
-            "Test suite left %d astray processes running. Killing those processes:\n%s",
-            len(children),
-            pprint.pformat(children),
-        )
-
-        _, alive = psutil.wait_procs(children, timeout=3, callback=on_terminate)
-        for child in alive:
-            try:
-                child.kill()
-            except psutil.NoSuchProcess:
-                continue
-
-        _, alive = psutil.wait_procs(alive, timeout=3, callback=on_terminate)
-        if alive:
-            # Give up
-            for child in alive:
-                log.warning(
-                    "Process %s survived SIGKILL, giving up:\n%s",
-                    child,
-                    pprint.pformat(child.as_dict()),
-                )
-
-
 @pytest.fixture(scope="session")
 def sminion():
     return create_sminion()
