+++ tests/pytests/integration/ssh/conftest.py
@@ -1,5 +1,7 @@
 import pytest
 
+from tests.support.pytest.helpers import reap_stray_processes
+
 
 @pytest.fixture(scope="package", autouse=True)
 def _auto_skip_on_fedora_39(grains):
@@ -11,3 +13,12 @@ def _auto_skip_on_fedora_39(grains):
             # has been deprecated since Python 3.7, so, the logic goes into trying to import
             # backports.ssl-match-hostname which is not installed on the system.
         )
+
+
+@pytest.fixture(autouse=True)
+def _reap_stray_processes(grains):
+    # when tests timeout, we migth leave child processes behind
+    # nuke them
+    with reap_stray_processes():
+        # Run test
+        yield
